refactor(platform): co-locate apiBaseUrlFromServerUrl with detectPlatform
Promote apiBaseUrlFromServerUrl() from pi-cli/src/octokit.ts into
pi-platform-github/src/provider.ts, alongside its natural pair
detectPlatform(). Both are pure functions of serverUrl with overlapping
host pattern-matching, answering complementary questions (platform type
vs API URL); keeping them split across packages was a drift risk
(adding a host meant editing both).

- pi-platform-github: add function + barrel export; move its tests into
  provider.spec.ts next to detectPlatform/isKnownServerUrl
- pi-cli: import from the barrel instead of defining locally; delete the
  now-empty octokit.spec.ts (its only coverage was the moved function)

No behavior change. Implements the pi-action-bridge constitution §12 Q9
prerequisite (Phase 1, item 1).
